Move JwtBuilder into Jwt folder
