+++ source/Messaging.PerformanceTest/MoveIn/MoveInService.cs
@@ -15,9 +15,8 @@ using System.Globalization;
 using System.Net.Http.Headers;
 using System.Security.Cryptography;
 using System.Text;
-using System.Text.Json;
 using System.Xml.Linq;
-using Messaging.PerformanceTest.Controllers;
+using Messaging.PerformanceTest.MoveIn.Jwt;
 
 namespace Messaging.PerformanceTest.MoveIn;
